sem_graph_conv updated to replace eye function

+++ models_baseline/gcn/sem_graph_conv.py
@@ -36,11 +36,11 @@ class SemGraphConv(nn.Module):
         h0 = torch.matmul(input, self.W[0])
         h1 = torch.matmul(input, self.W[1])
 
-        adj = -9e15 * torch.ones_like(self.adj).to(input.device)
+        adj = -9e15 * torch.ones_like(self.adj, dtype=torch.float32).to(input.device)
         adj[self.m] = self.e
         adj = F.softmax(adj, dim=1)
 
-        #M = torch.eye(adj.size(0), dtype=torch.float).to(input.device) # wally
+        #M_alt = torch.eye(adj.size(0), dtype=torch.float).to(input.device) # wally
         M = torch.from_numpy(np.array([[1., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
                                         [0., 1., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
                                         [0., 0., 1., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
